Factor out tkn_run.c
